MERGE 2.5.0-Pumpkin into master

+++ src/ggrc/gcalendar/calendar_event_builder.py
@@ -12,7 +12,7 @@ from ggrc import db
 from ggrc import settings
 from ggrc.models import all_models
 from ggrc.gcalendar import utils
-from ggrc.utils import benchmark
+from ggrc.utils import benchmark, generate_query_chunks
 
 
 logger = logging.getLogger(__name__)
@@ -31,9 +31,9 @@ class CalendarEventBuilder(object):
 
   def __init__(self):
     """Initialize CalendarEventBuilder."""
-    self.tasks = []
     self.events = []
     self.title_prefix = ""
+    self.chunk_size = 1000
     if settings.NOTIFICATION_PREFIX:
       self.title_prefix = "[{}] ".format(settings.NOTIFICATION_PREFIX)
 
@@ -49,7 +49,15 @@ class CalendarEventBuilder(object):
 
   def _preload_data(self):
     """Preload data for Calendar Event generation."""
-    self.tasks = all_models.CycleTaskGroupObjectTask.query.options(
+    self.events = all_models.CalendarEvent.query.all()
+
+  def _generate_events(self):
+    """Generates Calendar Events."""
+    task_mappings = utils.get_related_mapping(
+        left=all_models.CycleTaskGroupObjectTask,
+        right=all_models.CalendarEvent
+    )
+    columns = all_models.CycleTaskGroupObjectTask.query.options(
         orm.joinedload("cycle").load_only(
             "workflow_id",
             "is_current",
@@ -83,19 +91,18 @@ class CalendarEventBuilder(object):
             all_models.CycleTaskGroupObjectTask.title,
             all_models.CycleTaskGroupObjectTask.verified_date,
         ),
-    ).all()
-    self.events = all_models.CalendarEvent.query.all()
-
-  def _generate_events(self):
-    """Generates Calendar Events."""
-    task_mappings = utils.get_related_mapping(
-        left=all_models.CycleTaskGroupObjectTask,
-        right=all_models.CalendarEvent
-    )
-    for task in self.tasks:
-      events = task_mappings[task.id] if task.id in task_mappings else set()
-      self._generate_events_for_task(task, events_ids=events)
-    db.session.flush()
+    ).order_by(all_models.CycleTaskGroupObjectTask.end_date)
+    all_count = columns.count()
+    handled = 0
+    for query_chunk in generate_query_chunks(
+        columns, chunk_size=self.chunk_size, needs_ordering=False
+    ):
+      handled += query_chunk.count()
+      logger.info("Cycle task processed: %s/%s", handled, all_count)
+      for task in query_chunk:
+        events = task_mappings[task.id] if task.id in task_mappings else set()
+        self._generate_events_for_task(task, events_ids=events)
+      db.session.flush()
 
   def _generate_events_for_task(self, task, events_ids):
     """Generates CalendarEvents for CycleTaskGroupObjectTask."""
@@ -113,8 +120,9 @@ class CalendarEventBuilder(object):
             self._create_event_relationship(task, event)
             events_ids.discard(event.id)
 
-      for event_id in events_ids:
-        self._delete_event_relationship(event_id, task.id)
+      if self._should_delete_event_for(task):
+        for event_id in events_ids:
+          self._delete_event_relationship(event_id, task.id)
     except Exception as exp:   # pylint: disable=broad-except
       logger.exception("Generating of event for task %d has failed with the "
                        "following error %s.", task.id, exp.message)
@@ -202,6 +210,32 @@ class CalendarEventBuilder(object):
     ]
     return not any(conditions)
 
+  @staticmethod
+  def _should_delete_event_for(task):
+    """Determines should we delete a Calendar Event for the task or not.
+
+    Calendar events should NOT be deleted for:
+    - overdue cycle tasks
+    - finished cycle tasks
+    that does not satisfy any of these conditions:
+        - cycle task is deprecated.
+        - cycle task is verified (in case it has Verification flow).
+        - cycle task is finished (in case it has no Verification flow).
+        - cycle task is 'in progress' within a cycle that was ended
+          (tasks are stored at 'History' tab).
+        - cycle task is in archived workflow.
+    """
+    delete_conditions = [
+        task.status in [task.DEPRECATED, task.VERIFIED],
+        task.is_in_history,
+        task.workflow.workflow_archived,
+    ]
+    not_delete_conditions = [
+        task.status == task.FINISHED,
+        task.is_overdue,
+    ]
+    return not any(not_delete_conditions) or any(delete_conditions)
+
   def _generate_event_descriptions(self):
     """Generates CalendarEvents descriptions."""
     event_mappings = utils.get_related_mapping(
@@ -213,19 +247,32 @@ class CalendarEventBuilder(object):
             all_models.CalendarEvent.id,
             all_models.CalendarEvent.description,
         )
-    ).all()
-    for event in events:
-      if event.id not in event_mappings:
-        continue
-      self._generate_description_for_event(
-          event,
-          task_ids=event_mappings[event.id],
-      )
+    )
+
+    for query_chunk in generate_query_chunks(
+        events, chunk_size=self.chunk_size
+    ):
+      for event in query_chunk:
+        if event.id not in event_mappings:
+          continue
+        self._generate_description_for_event(
+            event,
+            task_ids=event_mappings[event.id],
+        )
 
   def _generate_description_for_event(self, event, task_ids):
     """Generates CalendarEvent descriptions based on tasks."""
-    titles = [u"- {}".format(unicode(task.title)) for task in self.tasks
-              if task.id in task_ids]
+    tasks = db.session.query(all_models.CycleTaskGroupObjectTask).filter(
+        all_models.CycleTaskGroupObjectTask.id.in_(task_ids)
+    ).order_by(all_models.CycleTaskGroupObjectTask.title).options(
+        load_only(
+            all_models.CycleTaskGroupObjectTask.id,
+            all_models.CycleTaskGroupObjectTask.title,
+        )
+    ).all()
+
+    titles = [u"- {}".format(unicode(task.title)) for task in tasks]
+
     event.description = (
         self.TASK_DESCRIPTION_HEADER +
         u"\n".join(titles) + u"\n" +
